(SERVER-2265) Refactor: renamed create
